fixed update feature

+++ subscriptionmanagerapp/src/app/components/employee/employee.component.ts
@@ -1,7 +1,7 @@
 import { HttpClient, HttpErrorResponse, HttpResponse } from '@angular/common/http';
 import { Component, OnInit } from '@angular/core';
 import { NgForm } from '@angular/forms';
-import {NgbModal, ModalDismissReasons} from '@ng-bootstrap/ng-bootstrap';
+import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { Employee } from 'src/app/model/employee';
 import { EmployeeService } from 'src/app/service/employee.service';
 
@@ -14,7 +14,7 @@ import { EmployeeService } from 'src/app/service/employee.service';
 })
 export class EmployeeComponent implements OnInit {
   public employees: Employee[] = [];
-  public selectedEmployee: Employee | undefined;
+  public selectedEmployee!: Employee; 
 
   constructor(private employeeService: EmployeeService, private modalService: NgbModal) { }
 
@@ -54,20 +54,26 @@ export class EmployeeComponent implements OnInit {
       })
   }
 
-  public deleteEmployeeById(id: number){
-    this.employeeService.deleteEmployeeById(id).subscribe(
-      (response: void)=>{
-        console.log("deleted " + response);
-        this.getEmployees();
-      },
-      (error : HttpErrorResponse)=>{
-        alert(error.message);
-      }
-    )
+  public deleteEmployeeById(id: string){
+    if(id !== this.selectedEmployee?.id){
+      //toast "they do not match"
+    }else{
+      this.employeeService.deleteEmployeeById(id).subscribe(
+        (response: void)=>{
+          console.log("deleted " + response);
+          this.getEmployees();
+        },
+        (error : HttpErrorResponse)=>{
+          alert(error.message);
+        }
+      )
+    }
   }
 
   openModal(employee: Employee, content: any){
     this.selectedEmployee = employee;
+    console.log(employee);
+    this.modalService.open(content);
   }
 
   public openAddModal(content: any){
